ceph_salt_deployment: avoid undersized PG warning
The ses7/octopus deployment recently started creating a pool
(device_health_metrics) with a single PG.

Before that, there were no pools created.

Now that we get a pool and PG, we are confronted with an undersized PG
warning. If the user has explicitly given "--single-node", let's assume
they want to avoid that warning.

Signed-off-by: Nathan Cutler <ncutler@suse.com>