contrib/standalone.sh: add release-testing script
contrib/standalone.sh is a script for finding regressions in sesdev code
changes. The idea is to run "sesdev create" with, and without,
--single-node for each of the main deployment versions:

- ses5
- ses6
- ses7
- nautilus
- octopus
- pacific

The idea is not to run the script interactively (though one could do
that), but rather to use GNU script, or some other suitable tool, to
capture the console log for later review.

I have placed the script in the "contrib" directory for now, to
emphasize the fact that it is still very much a work in progress.

Signed-off-by: Nathan Cutler <ncutler@suse.com>